refactor(config): move exception classes to an errors.py module
