+++ ggshield/core/config/__init__.py
@@ -1,4 +1,4 @@
-from .config import AccountConfig, Config, InstanceConfig, UnknownInstanceError
+from .config import AccountConfig, Config, InstanceConfig
 
 
-__all__ = ["AccountConfig", "Config", "InstanceConfig", "UnknownInstanceError"]
+__all__ = ["AccountConfig", "Config", "InstanceConfig"]
