+++ ggshield/core/config/config.py
@@ -5,6 +5,11 @@ from typing import Any, Dict, List, Mapping, Optional, Set, Tuple, Union
 
 import click
 
+from ggshield.core.config.errors import (
+    AuthExpiredError,
+    MissingTokenError,
+    UnknownInstanceError,
+)
 from ggshield.core.config.utils import (
     YAMLFileConfig,
     custom_asdict,
@@ -168,46 +173,6 @@ class InstanceConfig:
             self.account is not None
             and self.account.expire_at is not None
             and self.account.expire_at <= datetime.now(timezone.utc)
-        )
-
-
-class AuthError(click.ClickException):
-    """
-    Base exception for Auth-related configuration error
-    """
-
-    def __init__(self, instance: str, message: str):
-        super(AuthError, self).__init__(message)
-        self.instance = instance
-
-
-class UnknownInstanceError(AuthError):
-    """
-    Raised when the requested instance does not exist
-    """
-
-    def __init__(self, instance: str):
-        super(UnknownInstanceError, self).__init__(
-            instance, f"Unknown instance: '{instance}'"
-        )
-
-
-class AuthExpiredError(AuthError):
-    """
-    Raised when authentication has expired for the given instance
-    """
-
-    def __init__(self, instance: str):
-        super(AuthExpiredError, self).__init__(
-            instance,
-            f"Instance '{instance}' authentication expired, please authenticate again.",
-        )
-
-
-class MissingTokenError(AuthError):
-    def __init__(self, instance: str):
-        super(MissingTokenError, self).__init__(
-            instance, f"No token is saved for this instance: '{instance}'"
         )
